+++ include/utils/sensors.h
@@ -1,4 +1,4 @@
-/// Copyright (C) 2021 Christian Brommer and Alessandro Fornasier,
+/// Copyright (C) 2021 Martin Scheiber, Alessandro Fornasier, and others,
 /// Control of Networked Systems, Universitaet Klagenfurt, Austria
 ///
 /// All rights reserved.
@@ -7,9 +7,8 @@
 /// no commercial use allowed, the full terms of which are made available
 /// in the LICENSE file. No license in patents is granted.
 ///
-/// You can contact the authors at <christian.brommer@ieee.org>
+/// You can contact the authors at <martin.scheiber@ieee.org>
 /// and <alessandro.fornasier@ieee.org>
-
 
 #ifndef SENSORS_H
 #define SENSORS_H
@@ -21,26 +20,35 @@ namespace utils
 {
   namespace sensors
   {
-    /**
-     * @brief Struct for imu data
-     */
-    struct imuData
+    ///
+    /// \brief The sensorData struct is a interface for sensor data structures.
+    /// \note This structure only implements operators for comparision reasons
+    ///
+    struct sensorData
     {
+      double timestamp;
 
-        /// Timestamp of the reading (s)
-        double timestamp;
+      /// Sort function to allow for using of STL containers
+      bool operator<(const sensorData& other) const {
+          return timestamp < other.timestamp;
+      }
+    };
 
-        /// Gyroscope reading, angular velocity (rad/s)
-        Eigen::Matrix<double, 3, 1> wm;
+    ///
+    /// \brief The imuData struct
+    ///
+    struct imuData : sensorData
+    {
+        Eigen::Matrix<double, 3, 1> wm;     //!<  Gyroscope reading, angular velocity (rad/s)
+        Eigen::Matrix<double, 3, 1> am;     //!< Accelerometer reading, linear acceleration (m/s^2)
+    };
 
-        /// Accelerometer reading, linear acceleration (m/s^2)
-        Eigen::Matrix<double, 3, 1> am;
-
-        /// Sort function to allow for using of STL containers
-        bool operator<(const imuData& other) const {
-            return timestamp < other.timestamp;
-        }
-
+    ///
+    /// \brief The lrfData struct
+    ///
+    struct lrfData : sensorData
+    {
+      double range;                         //!< Range reading, distance (m)
     };
   }
 }